[PyCOMPSs] Jupyter notebook clean temp files after stop.
git-svn-id: http://compss.bsc.es/svn/compss/framework/trunk@3096 9ab3861e-6c05-4e1b-b5ef-99af60850597

Former-commit-id: 934dd6bcde356d1fa5907eb99451c1d92a7686f2

+++ compss/programming_model/bindings/python/src/pycompss/interactive.py
@@ -335,7 +335,10 @@ def cleanTempFiles():
     Currently: app_path, which contains the file path where all interactive code required by the worker is.
     '''
     try:
-        os.remove(app_path)
+        if os.path.exists(app_path):
+            os.remove(app_path)
+        if os.path.exists(app_path + 'c'):
+            os.remove(app_path + 'c')
     except OSError:
         print "[ERROR] An error has occurred when cleaning temporary files."
